requirements.txt
django

django-debug-toolbar

django-phonenumber-field

phonenumbers

django-jsignature

django-filter

django_reverse_admin
